bugfix db connections
