+++ prediction-polls/backend/src/repositories/PollDB.js
@@ -1,18 +1,10 @@
-const mysql = require('mysql2');
 const { addRefreshToken, deleteRefreshToken, findUser } = require('./AuthorizationDB');
 const { updatePoints } = require('./ProfileDB');
 const moment = require('moment');
 const errorCodes = require("../errorCodes.js");
+const {pool} = require("./DatabaseInit.js")
 
 require('dotenv').config();
-
-const pool = mysql.createPool({
-    host: process.env.MYSQL_HOST,
-    user: process.env.MYSQL_USER,
-    password: process.env.MYSQL_PASSWORD,
-    database: process.env.MYSQL_DATABASE
-}).promise()
-
 
 async function getPolls(){
     const sql = 'SELECT * FROM polls';
