+++ prediction-polls/backend/src/repositories/ProfileDB.js
@@ -1,16 +1,8 @@
 const authDb = require("../repositories/AuthorizationDB.js");
-const mysql = require('mysql2')
 const errorCodes = require("../errorCodes.js")
+const {pool} = require("./DatabaseInit.js")
 
 require('dotenv').config();
-
-const pool = mysql.createPool({
-    host: process.env.MYSQL_HOST,
-    user: process.env.MYSQL_USER,
-    password: process.env.MYSQL_PASSWORD,
-    database: process.env.MYSQL_DATABASE
-}).promise()
-
 
 async function getProfileWithProfileId(profileId) {
