+++ prediction-polls/backend/src/repositories/SemanticDB.js
@@ -1,16 +1,7 @@
-
-const mysql = require('mysql2')
 const errorCodes = require("../errorCodes.js")
+const {pool} = require("./DatabaseInit.js")
 
 require('dotenv').config();
-
-const pool = mysql.createPool({
-    host: process.env.MYSQL_HOST,
-    user: process.env.MYSQL_USER,
-    password: process.env.MYSQL_PASSWORD,
-    database: process.env.MYSQL_DATABASE
-}).promise()
-
 
 async function getSemanticTagsOfPoll(poll_id) {
     try {
